Adds support for --files0-from=F
A good amount of things had to be reworked to add support for this
feature. Most notably, most of the methods of the options class have
been changed from void return types to return a new type of enum
"OptionParseType". Now a return code is returned for each argument,
which also makes it less hacky to handle toggling default options.