Clean up Lexer and Jason
This commit does not add any new functionality, rather it removes a lot
of the debug print lines that were left in Lexer and Jason (although not
all). It also breaks up Jason::processTokens() into multiple private
helper functions. The functionality is the same, but now processTokens
is less than 300 lines long.