WIP: Add Jason print method
Adds the Jason print method, which is currently buggy and nonfunctional.
There appears to be a problem with either the generation or printing of
nested objects. It's hard to debug because printing would help me see if
the generation is broken, but alas, here we are. Committed in a broken
state because I got tired of working on this today.